% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{EUROSTAT_import_data_points}
\alias{EUROSTAT_import_data_points}
\title{Insert data points from BS}
\usage{
EUROSTAT_import_data_points(code, con, schema = "platform", use_cache = FALSE)
}
\arguments{
\item{code}{Eurostat code name of the table}

\item{con}{Database connection}

\item{schema}{Schema name}

\item{use_cache}{logical, set to true for testing, false in production}
}
\value{
Insertion results (invisibly)
}
\description{
Function to prepare and insert BS data points. The function first prepares
the required vintages and inserts them, then prepares the data points
table and inserts it. The function returns the results invisibly.
}
\details{
This is a BS specific function, which should be followed by the generic
UMARimportR function to write the vintage hashes and clean up redundant
vintages.
}
