% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{EUROSTAT_import_structure}
\alias{EUROSTAT_import_structure}
\title{Import structure for a single dataset (categories, relationships, and table info)}
\usage{
EUROSTAT_import_structure(
  con,
  dataset_code,
  source_id = 7,
  schema = "platform"
)
}
\arguments{
\item{con}{Database connection object}

\item{dataset_code}{Character, Eurostat dataset code}

\item{source_id}{Integer source ID, default 7 for Eurostat}

\item{schema}{defautls to "platform"}
}
\description{
Import structure for a single dataset (categories, relationships, and table info)
}
