% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eurostat_unit_map}
\alias{eurostat_unit_map}
\title{Dataset: eurostat_unit_map}
\format{
A data frame with [193] rows and [16] variables:
\describe{
  \item{eurostat_unit}{unit code on Eurostat}
  \item{unar_unit}{unit code in our database}
}
}
\usage{
eurostat_unit_map
}
\description{
Mapping of Eurostat units and indicators to our internal
database unit list
}
\keyword{datasets}
