% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{format_period_id}
\alias{format_period_id}
\title{Helper function to format period IDs}
\usage{
format_period_id(dates, interval)
}
\arguments{
\item{dates}{date}

\item{interval}{M or Q or A}
}
\description{
converts date to period id. (originally from DESEZONIRANJEfetchR)
}
\keyword{internal}
