% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy_parsing_helpers.R
\name{get_dataset_ancestors}
\alias{get_dataset_ancestors}
\title{Get ancestor folders for a dataset from TOC}
\usage{
get_dataset_ancestors(dataset_code, toc)
}
\arguments{
\item{dataset_code}{Character, Eurostat dataset code}

\item{toc}{Data frame, output from eurostat::get_eurostat_toc()}
}
\value{
Data frame with columns: title, hierarchy, parent_title, instance
}
\description{
Get ancestor folders for a dataset from TOC
}
