% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy_parsing_helpers.R
\name{get_next_category_id}
\alias{get_next_category_id}
\title{Get next available category ID for a source}
\usage{
get_next_category_id(con, source_id)
}
\arguments{
\item{con}{Database connection object}

\item{source_id}{Integer source ID}
}
\value{
Integer, next available category ID
}
\description{
Get next available category ID for a source
}
