% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_category_table_table}
\alias{prepare_category_table_table}
\title{Prepare table to insert into `category_table` table}
\usage{
prepare_category_table_table(
  code,
  toc,
  con,
  source_id = 7,
  schema = "platform"
)
}
\arguments{
\item{code}{the original Eurostat code (e.g. agr_r_animal)}

\item{toc}{output from eurostat::get_eurostat_toc()}

\item{con}{a connection to the database}

\item{source_id}{integer, the source ID (default 7 for Eurostat)}

\item{schema}{the schema to use for the connection, default is "platform"}
}
\value{
a dataframe with the `category_id` `table_id` and `source_id` columns for
each table-category relationship.
}
\description{
Helper function that extracts the parent category for each table from the full
hierarchy toc, and fills up the category_table table with the table ids and
their categories (parents). A single table can have multiple parents - meaning
it is member of several categories. Returns table
ready to insert into the `category_table` table with the db_writing family of functions.
}
