% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_dimension_levels_table}
\alias{prepare_dimension_levels_table}
\title{Prepare table to insert into `dimension_levels` table}
\usage{
prepare_dimension_levels_table(
  code,
  dim_structure = NULL,
  con,
  schema = "platform"
)
}
\arguments{
\item{code}{the original Eurostat code (e.g. agr_r_animal)}

\item{dim_structure}{output from extract_dimension_structure() (optional, will be computed if not provided)}

\item{con}{a connection to the database}

\item{schema}{the schema to use for the connection, default is "platform"}
}
\value{
a dataframe with `tab_dim_id`, `level_value`, and `level_text` columns
}
\description{
Extracts dimension levels from a Eurostat dataset and prepares them for
insertion into the dimension_levels table.
}
