% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure.R
\name{prepare_table_table}
\alias{prepare_table_table}
\title{Prepare table to insert into `table` table}
\usage{
prepare_table_table(code, toc, con, schema = "platform", keep_vintage = FALSE)
}
\arguments{
\item{code}{the original Eurostat code (e.g. agr_r_animal)}

\item{toc}{output from eurostat::get_eurostat_toc()}

\item{con}{a connection to the database}

\item{schema}{the schema to use for the connection, default is "platform"}

\item{keep_vintage}{boolean whether to keep vintages}
}
\value{
a dataframe with the `code`, `name`, `source_id`, `url`, and `notes` columns
for this table.
}
\description{
Returns table ready to insert into the `table` table with the db_writing family
of functions.
}
