% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toc}
\alias{toc}
\title{Dataset: toc}
\format{
A data frame with [11882] rows and [9] variables:
\describe{
  \item{title}{title}
  \item{code}{code}
  \item{type}{type}
  \item{last.update.of.data}{last.update.of.data}
  \item{last.table.structure.change}{last.table.structure.change}
  \item{data.start}{data.start}
  \item{data.end}{data.end}
  \item{values}{values}
  \item{hierarchy}{hierarchy}
}
}
\usage{
toc
}
\description{
eurosat toc
}
\keyword{datasets}
