% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{EUROSTAT_import_structure}
\alias{EUROSTAT_import_structure}
\title{Import structure for a single dataset (categories, relationships, and table info)}
\usage{
EUROSTAT_import_structure(
  con,
  code,
  source_id = 7,
  schema = "platform",
  all_levels = FALSE,
  keep_vintage = FALSE,
  use_cache = FALSE
)
}
\arguments{
\item{con}{Database connection object}

\item{code}{Character, Eurostat dataset code}

\item{source_id}{Integer source ID, default 7 for Eurostat}

\item{schema}{defautls to "platform"}

\item{all_levels}{logical whether to let user use dimension selector, default
is FALSE, which automatically selects all levels in all dimensions}

\item{keep_vintage}{logical indicating whether to keep vintages, defaults to F}

\item{use_cache}{logical, set to true for testing, false in production}
}
\description{
Import structure for a single dataset (categories, relationships, and table info)
}
